requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.3
brotli==1.2.0
celery==5.5.3
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.7.0
django==4.2.26
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
idna==3.11
kombu==5.5.4
packaging==25.0
pillow==10.4.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==14.0.1
tinycss2==1.4.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.7.0
xlsxwriter==3.2.9
zopfli==0.2.3.post1
